+++ src/lib/constants.ts
@@ -1,3 +1,17 @@
 export const MINIMUM_SEASON_THRESHOLD = 70;
 export const MAXIMUM_GUILD_MEMBERS = 30;
 export const MAXIMUM_TOKENS_PER_SEASON = 29;
+
+export const BOSS_EMOJIS: Record<string, string> = {
+    Szarekh: "<:Szarekh:1385343132950069278>",
+    TyrantLeviathan: "<:TyrantLeviathan:1385342042170851334>",
+    TyrantGorgon: "<:TyrantGorgon:1385340907351441619>",
+    TyrantKronos: "<:TyrantKronos:1385341128626409522>",
+    Ghazghkull: "<:Ghazghkull:1385340195494170664>",
+    Avatar: "<:Avatar:1385338950834716802>",
+    Magnus: "<:Magnus:1385342412217520379>",
+    Mortarion: "<:Mortarion:1385342557969453197>",
+    Belisarius: "<:Cawl:1385339595578806312>",
+    RogalDornTank: "<:RogalDornTank:1385342727037784174>",
+    Screamer: "<:ScreamerKiller:1385342920302788608>",
+};